feat: adiciona serviços de Gemini e SlideStories
