+++ src/services/gemini-client.js
@@ -1,0 +1,10 @@
+import { text } from './gemini';
+
+async function generateContent() {
+  console.log(text);
+
+  const pagePrompt = document.querySelector('.prompt');
+  pagePrompt.innerHTML = text;
+}
+
+generateContent();
