+++ src/services/gemini.js
@@ -1,20 +1,24 @@
-import { GoogleGenerativeAI } from "@google/generative-ai";
+import { GoogleGenerativeAI } from '@google/generative-ai';
 
 const API_KEY = process.env.KEY_GOOGLE_AI_STUDIO;
 
 const genAI = new GoogleGenerativeAI(API_KEY);
 
-const model = genAI.getGenerativeModel({ model: "gemini-pro" });
+const model = genAI.getGenerativeModel({ model: 'gemini-pro' });
 
-const prompt = "fale sobre girafas em 5 palavras";
+const prompt = 'fale sobre girafas em 5 palavras';
 
-async function generateContent() {
-  const result = await model.generateContent(prompt);
-  const response = result.response;
-  const text = response.text();
+let text = '';
 
-  const pagePrompt = document.querySelector(".prompt");
-  pagePrompt.innerHTML = text;
-}
+model
+  .generateContent(prompt)
+  .then((result) => {
+    const response = result.response;
+    text = response.text();
+    console.log(text);
+  })
+  .catch((error) => {
+    console.error(error);
+  });
 
-generateContent();
+export { text };
